Perform Reddit fetch/analysis on a helper thread
These tasks can take several seconds in a good case. Worst case tests in
the past have taken several minutes to complete.

I don't want to risk having other readers waiting for this to complete or
requiring block_on_read=off to be set to support multiple simultaneous
users, so this has to be done on a separate thread - just like database
searching/Telnet in cadence-server.