Show error box on parsing error
Instead of appearing to mysteriously stop if the server returns data we're
not expecting, we now display the generic error box.

This is a UX issue I noticed while testing that the server returned the
data I expected it to - The UI hid the loading animation, but failed to
load in results (as expected, since the results aren't the same as what
the client was written to read), displaying no feedback whatsoever.

Now, it should display the generic "something went wrong" box.